+++ back-end/src/main/java/br/com/sistema/Service/ProfessorService.java
@@ -9,6 +9,7 @@ import br.com.sistema.Mapper.ProfessorMapper;
 import br.com.sistema.Model.Curso;
 import br.com.sistema.Model.Professor;
 import br.com.sistema.Repository.AlunoRepository;
+import br.com.sistema.Repository.AulaRepository;
 import br.com.sistema.Repository.CursoRepository;
 import br.com.sistema.Repository.ProfessorRepository;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +24,7 @@ public class ProfessorService {
     private final ProfessorRepository repository;
     private final ProfessorMapper mapper;
 
+    private final AulaRepository aulaRepository;
     private final AlunoRepository alunoRepository;
 
     public ProfessorDTO create(ProfessorDTO professorDTO){
@@ -43,15 +45,19 @@ public class ProfessorService {
 
     public ProfessorDTO update(ProfessorDTO professorDTO, Long id){
 
-        if (repository.existsByMatricula(professorDTO.getMatricula())) {
-            throw new BusinessException("Já existe um professor com a matrícula '" + professorDTO.getMatricula() + "'.");
+        ProfessorDTO old = findById(id);
+
+
+        if(!professorDTO.getMatricula().equals(old.getMatricula())){
+            if (repository.existsByMatricula(professorDTO.getMatricula())) {
+                throw new BusinessException("Já existe um professor com a matrícula '" + professorDTO.getMatricula() + "'.");
+            }
+
+            if (alunoRepository.existsByMatricula(professorDTO.getMatricula())) {
+                throw new EntityNotFoundException("Já existe um aluno com a matrícula '" + professorDTO.getMatricula() + "'.");
+            }
         }
 
-        if (alunoRepository.existsByMatricula(professorDTO.getMatricula())) {
-            throw new EntityNotFoundException("Já existe um aluno com a matrícula '" + professorDTO.getMatricula() + "'.");
-        }
-
-        findById(id);
 
         Professor entity = mapper.toEntity(professorDTO);
         entity.setId(id);
@@ -63,6 +69,10 @@ public class ProfessorService {
     public void delete(Long id) {
         Professor professor = repository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Professor com ID '" + id + "' não encontrado."));
+
+        if(aulaRepository.existsAulasByProfessorId(professor.getId())){
+            throw new BusinessException("Este professor possui aulas cadastradas e não pode ser excluido.");
+        }
 
         repository.delete(professor);
     }
